refactor: split handleSlashCommand into per-domain handlers
The chat-TUI slash dispatcher had grown past the file-length smell. Move the
domain handlers into commands/handlers/ (agent/model, api-mode, subscription,
login, approval, memory) plus a shared slashContext, leaving
handleSlashCommand.ts as just the dispatch switch. Public symbols used by
runChatTui and the slash tests are re-exported from handleSlashCommand, so no
importer changes. Pure refactor — no behavior change.
