+++ packages/cli/src/chat/tui/commands/handleSlashCommand.ts
@@ -1,480 +1,61 @@
-import { getAgent } from '@agent/index';
-import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import { executionRegistry } from '@agent/runtime/executionRegistry';
 import { notifyFollowUpSent } from '@agent/toolUse/ToolUseFollowUp';
 import { ToolUseFollowUpQueue } from '@agent/toolUse/ToolUseFollowUpQueueManager';
-import {
-  CODEX_PREFER_SUBSCRIPTION_KEY,
-  getCodexStatus,
-  isCodexSignedIn,
-  isPreferCodexSubscription,
-} from '@auth/codex';
-import { invalidateModelOptionsCache } from '@model/computeModelOptions';
-import { platform } from '@platform/platform';
-import {
-  formatCliApiMode,
-  parseCliApiMode,
-  setCliApiMode,
-  type CliApiMode,
-} from '@cli/runtime/apiAccessMode';
-import { loadCliApiStatusLines } from '@cli/runtime/apiStatus';
-import { assertCliAgentLaunch } from '@cli/runtime/agents';
-import { type CliContext, CliUsageError } from '@cli/runtime/cliContext';
-import { setCliHelperModel } from '@cli/runtime/initPlatform';
-import {
-  githubSelectAccountWarning,
-  parseChatLoginSlashArgs,
-  type CliLoginSlashArgs,
-} from '@cli/runtime/loginOptions';
-import {
-  formatCliMemoryList,
-  formatCliMemoryPreview,
-} from '@cli/runtime/memory';
-import {
-  formatCliNoAvailableModelsRecovery,
-  type CliNoAvailableModelsRecoveryOptions,
-} from '@cli/runtime/modelAccess';
-import { selectCliRootModel } from '@cli/runtime/rootModelSelection';
-import {
-  formatCliApprovalPolicy,
-  parseCliApprovalPolicy,
-} from '@cli/runtime/approvalPolicyText';
+import { formatCliApiMode } from '@cli/runtime/apiAccessMode';
+import { formatCliApprovalPolicy } from '@cli/runtime/approvalPolicyText';
 import { parseCliHistoryId } from '@cli/runtime/history';
 import { defaultShortcutModifierLabel } from '@cli/runtime/shortcutLabels';
-import {
-  formatCliManualAuthUrlMessage,
-  relayTokenStillActiveNotice,
-  signInCliSupabase,
-  signInCliSupabaseDeviceCode,
-  signOutCliSupabase,
-} from '@cli/runtime/supabaseAuth';
-import { formatCliDeviceAuthMessage } from '@cli/runtime/supabaseAuthDeviceCode';
-import type { CliApprovalPolicy } from '@cli/schemas/cliSettings';
 import { toErrorMessage } from '@common/errors/errorMessage';
-import { type ExecutionId } from '@shared/schemas';
-import { DELEGATION_TOOLS } from '@shared/constants/delegationTools';
 import { GoalStore } from '@tools/goal';
-import { loadMemoryItems } from '@tools/memory/memoryFileSystem';
 
 import { formatCliSessionStatus } from '../sessionStatus';
 import { requestCliCompaction } from '../state/compactionRequest';
-import { cliState, setCliSessionModelOverride } from '../state/cliState';
-import {
-  chatTuiCanStartRootRun,
-  type TuiSession,
-} from '../state/sessionRunState';
+import { cliState } from '../state/cliState';
+import { chatTuiCanStartRootRun } from '../state/sessionRunState';
 import { terminalCapabilities } from '../state/terminalCapabilities';
 import {
   appendLocalAssistantTranscript,
   appendLocalUserTranscript,
 } from '../state/transcript';
 import { formatSlashCommandHelp, GOAL_MODE_HELP } from './helpText';
+import { applyInitialCliAgentSelection } from './handlers/agentModelCommands';
 import {
-  openCliSlashCommandForm,
-  openRegisteredCliSlashForm,
-} from './slashForms';
+  applyCliApiModeSelection,
+  showCliAuthStatus,
+} from './handlers/apiModeCommands';
+import { applyCliApprovalPolicySelection } from './handlers/approvalCommand';
+import { loginFromChat, logoutFromChat } from './handlers/loginCommands';
+import {
+  showCliMemoryList,
+  showCliMemoryPreview,
+} from './handlers/memoryCommands';
+import {
+  openCanonicalSlashForm,
+  type SlashCommandContext,
+} from './handlers/slashContext';
+import { applyCliSubscriptionToggle } from './handlers/subscriptionCommand';
+import { openRegisteredCliSlashForm } from './slashForms';
 import {
   findSlashCommand,
   listSlashCommands,
   parseSlashInput,
   suggestSlashCommand,
-  type SlashCommand,
 } from './slashRegistry';
 
-export interface SlashCommandContext {
-  readonly session: TuiSession;
-  readonly commandName?: string;
-  readonly cwd: CliContext['cwd'];
-  readonly processCwd?: CliContext['cwd'];
-  readonly initialAgent: string;
-  readonly initialModel: string;
-  readonly interruptActive: () => void;
-  readonly requestInputExit: () => void;
-  readonly getApprovalPolicy: () => CliApprovalPolicy;
-  readonly setApprovalPolicy: (policy: CliApprovalPolicy) => void;
-  readonly canSelectModel: () => boolean;
-  readonly resetSession: () => void;
-  readonly resumeExecution: (id: ExecutionId) => Promise<void>;
-}
-
-export const CHAT_API_MODE_MODEL_RECOVERY = {
-  includedModeAction: 'switch to included relay with `/api included`',
-  loginAction: 'run `/login`',
-  personalModeAction: 'switch to personal API keys with `/api personal`',
-  configureKeyAction: 'configure a provider API key',
-} satisfies CliNoAvailableModelsRecoveryOptions;
-
-export function chatAgentSupportsDelegation(agentName: string): boolean {
-  return (
-    getAgent(agentName, AgentCategory.ToolUse)?.tools?.some((toolName) =>
-      DELEGATION_TOOLS.has(toolName),
-    ) ?? false
-  );
-}
-
-export function chatToolUseAgentUsageError(
-  agentName: string,
-): string | undefined {
-  try {
-    assertCliAgentLaunch(
-      agentName,
-      getAgent(agentName, AgentCategory.ToolUse),
-      'chat',
-    );
-    return undefined;
-  } catch (error) {
-    if (error instanceof CliUsageError) return error.message;
-    throw error;
-  }
-}
-
-export function applyInitialCliAgentSelection(
-  agentName: string,
-  context: SlashCommandContext,
-): void {
-  if (!chatTuiCanStartRootRun(context.session)) {
-    appendLocalAssistantTranscript(
-      'Agent changes are only available before the first message. Start a new chat with texra chat --agent=<name> to choose a different root agent.',
-    );
-    return;
-  }
-
-  const nextAgent = agentName.trim();
-  const usageError = chatToolUseAgentUsageError(nextAgent);
-  if (usageError) {
-    appendLocalAssistantTranscript(usageError);
-    return;
-  }
-  cliState.sessionMeta.set({
-    ...cliState.sessionMeta.get(),
-    agent: nextAgent,
-    canDelegate: chatAgentSupportsDelegation(nextAgent),
-  });
-  appendLocalAssistantTranscript(`Root agent set to ${nextAgent}.`);
-}
-
-export async function applyCliModelSelection(
-  model: string,
-  context: SlashCommandContext,
-): Promise<void> {
-  const nextModel = model.trim();
-  if (chatTuiCanStartRootRun(context.session)) {
-    try {
-      const { apiMode } = cliState.sessionMeta.get();
-      const selection = await selectCliRootModel({
-        model: nextModel,
-        modelSource: 'override',
-        apiMode,
-        noAvailableModelsMessage: formatCliNoAvailableModelsRecovery(
-          apiMode,
-          CHAT_API_MODE_MODEL_RECOVERY,
-        ),
-      });
-      setCliSessionModelOverride(selection.model);
-      appendLocalAssistantTranscript(`Root model set to ${selection.model}.`);
-    } catch (error: unknown) {
-      appendLocalAssistantTranscript(toErrorMessage(error));
-    }
-    return;
-  }
-
-  if (!context.canSelectModel()) {
-    appendLocalAssistantTranscript(
-      'Finish the active response before switching models.',
-    );
-    return;
-  }
-
-  const activeFlow = context.session.streamId
-    ? executionRegistry.getToolUseFlowContext(context.session.streamId)
-    : undefined;
-  if (!activeFlow) {
-    appendLocalAssistantTranscript(
-      'Model switching is only available for an active tool-use chat. Start a new chat with texra chat --model=<name> to choose a different root model.',
-    );
-    return;
-  }
-
-  try {
-    await activeFlow.switchModel(nextModel);
-    setCliSessionModelOverride(nextModel);
-  } catch (error: unknown) {
-    appendLocalAssistantTranscript(toErrorMessage(error));
-    return;
-  }
-
-  try {
-    await setCliHelperModel(nextModel);
-  } catch (error: unknown) {
-    appendLocalAssistantTranscript(
-      `Model switched to ${nextModel}. Could not persist it as the default helper model: ${toErrorMessage(error)}`,
-    );
-    return;
-  }
-
-  appendLocalAssistantTranscript(
-    `Model switched to ${nextModel}. Future turns will use it.`,
-  );
-}
-
-async function reconcileRootModelAfterApiModeChange(
-  context: SlashCommandContext | undefined,
-  apiMode: CliApiMode,
-): Promise<string | undefined> {
-  if (!context || !chatTuiCanStartRootRun(context.session)) return undefined;
-
-  const { model: currentModel, modelSource } = cliState.sessionMeta.get();
-  const selection = await selectCliRootModel({
-    model: currentModel,
-    modelSource,
-    apiMode,
-    noAvailableModelsMessage: formatCliNoAvailableModelsRecovery(
-      apiMode,
-      CHAT_API_MODE_MODEL_RECOVERY,
-    ),
-  });
-  if (selection.model === currentModel) return undefined;
-
-  cliState.sessionMeta.set({
-    ...cliState.sessionMeta.get(),
-    model: selection.model,
-  });
-  return selection.notice;
-}
-
-function setCliSessionApiMode(apiMode: CliApiMode): void {
-  cliState.sessionMeta.set({
-    ...cliState.sessionMeta.get(),
-    apiMode,
-  });
-}
-
-export async function applyCliApiModeSelection(
-  mode: string | CliApiMode,
-  context?: SlashCommandContext,
-): Promise<void> {
-  const normalized = mode.trim().toLowerCase();
-
-  if (!normalized || normalized === 'status') {
-    const lines = await loadCliApiStatusLines({
-      apiMode: cliState.sessionMeta.get().apiMode,
-    });
-    appendLocalAssistantTranscript(
-      [...lines, 'Usage: /api personal | /api included'].join('\n'),
-    );
-    return;
-  }
-
-  const apiMode = parseCliApiMode(normalized);
-  if (apiMode) {
-    await setCliApiMode(apiMode);
-    setCliSessionApiMode(apiMode);
-    let modelNotice: string | undefined;
-    try {
-      modelNotice = await reconcileRootModelAfterApiModeChange(
-        context,
-        apiMode,
-      );
-    } catch (error: unknown) {
-      modelNotice = toErrorMessage(error);
-    }
-    appendLocalAssistantTranscript(
-      [
-        `API mode set to ${formatCliApiMode(apiMode)}.`,
-        ...(modelNotice ? [modelNotice] : []),
-      ].join('\n'),
-    );
-    return;
-  }
-
-  appendLocalAssistantTranscript('Usage: /api personal | /api included');
-}
-
-const SUBSCRIPTION_ON = new Set(['on', 'enable', 'enabled', 'true', 'yes']);
-const SUBSCRIPTION_OFF = new Set(['off', 'disable', 'disabled', 'false', 'no']);
-const SUBSCRIPTION_USAGE = 'Usage: /subscription on | off';
-
-/**
- * Toggle "prefer ChatGPT subscription" for Codex-eligible models. The
- * preference is orthogonal to the relay/personal api-mode (it only changes
- * where Codex models get their credentials), so it lives in its own command
- * rather than the `/api` binary.
- */
-async function applyCliSubscriptionToggle(input: string): Promise<void> {
-  const normalized = input.trim().toLowerCase();
-  const status = await getCodexStatus();
-  const accountLine = status.signedIn
-    ? `Signed in with ChatGPT as ${status.email ?? status.accountId ?? 'your account'}.`
-    : 'Not signed in with ChatGPT — run `texra auth chatgpt login`.';
-
-  if (!normalized || normalized === 'status') {
-    appendLocalAssistantTranscript(
-      [
-        `ChatGPT subscription: ${isPreferCodexSubscription() ? 'on' : 'off'}.`,
-        accountLine,
-        SUBSCRIPTION_USAGE,
-      ].join('\n'),
-    );
-    return;
-  }
-
-  let enabled: boolean;
-  if (SUBSCRIPTION_ON.has(normalized)) enabled = true;
-  else if (SUBSCRIPTION_OFF.has(normalized)) enabled = false;
-  else {
-    appendLocalAssistantTranscript(SUBSCRIPTION_USAGE);
-    return;
-  }
-
-  await platform().config.update(
-    CODEX_PREFER_SUBSCRIPTION_KEY,
-    enabled,
-    'global',
-  );
-  invalidateModelOptionsCache();
-  // Prime the synchronous sign-in hint so routing reflects the change at once.
-  await isCodexSignedIn();
-
-  const lines = [
-    `ChatGPT subscription ${enabled ? 'enabled' : 'disabled'} for Codex models.`,
-  ];
-  if (enabled && !status.signedIn) {
-    lines.push(
-      'You are not signed in with ChatGPT yet — run `texra auth chatgpt login` to use it.',
-    );
-  } else if (enabled) {
-    lines.push(accountLine);
-  }
-  appendLocalAssistantTranscript(lines.join('\n'));
-}
-
-async function showCliAuthStatus(): Promise<void> {
-  appendLocalAssistantTranscript(
-    (
-      await loadCliApiStatusLines({
-        apiMode: cliState.sessionMeta.get().apiMode,
-      })
-    ).join('\n'),
-  );
-}
-
-const CHAT_LOGIN_USAGE =
-  'Usage: /login [github | google] [--no-browser] [--device] [--select-account] [--login-hint <account>]';
-
-function loginStartMessage(args: CliLoginSlashArgs): string {
-  if (args.device) return 'Starting TeXRA device-code sign-in.';
-  if (args.noBrowser) return `Starting TeXRA ${args.provider} sign-in.`;
-  return `Opening browser for TeXRA ${args.provider} sign-in...`;
-}
-
-async function loginFromChat(input: string): Promise<void> {
-  const args = parseChatLoginSlashArgs(input);
-  if (!args) {
-    appendLocalAssistantTranscript(CHAT_LOGIN_USAGE);
-    return;
-  }
-
-  const accountWarning = githubSelectAccountWarning(args);
-  if (accountWarning) appendLocalAssistantTranscript(accountWarning);
-  appendLocalAssistantTranscript(loginStartMessage(args));
-
-  try {
-    const session = args.device
-      ? await signInCliSupabaseDeviceCode({
-          onDeviceCode: (authorization) => {
-            appendLocalAssistantTranscript(
-              formatCliDeviceAuthMessage(authorization),
-            );
-          },
-        })
-      : await signInCliSupabase({
-          provider: args.provider,
-          openBrowser: !args.noBrowser,
-          selectAccount: args.selectAccount,
-          loginHint: args.loginHint,
-          manualBrowserHint: '/login --no-browser',
-          onAuthUrl: (url) => {
-            if (args.noBrowser) {
-              appendLocalAssistantTranscript(
-                formatCliManualAuthUrlMessage(url),
-              );
-            }
-          },
-        });
-    setCliSessionApiMode('included');
-    appendLocalAssistantTranscript(
-      [
-        `Signed in as ${session.account.label}.`,
-        ...(await loadCliApiStatusLines({ apiMode: 'included' })),
-      ].join('\n'),
-    );
-  } catch (error: unknown) {
-    appendLocalAssistantTranscript(toErrorMessage(error));
-  }
-}
-
-async function logoutFromChat(): Promise<void> {
-  try {
-    await signOutCliSupabase();
-    // Sign-out only clears the stored session; a configured TEXRA_RELAY_TOKEN
-    // keeps authenticating relay calls, so report it — and keep the session
-    // in included mode so the notice matches what actually happens.
-    const relayNotice = relayTokenStillActiveNotice();
-    const apiMode = relayNotice ? 'included' : 'personal';
-    setCliSessionApiMode(apiMode);
-    appendLocalAssistantTranscript(
-      [
-        'Signed out.',
-        ...(relayNotice ? [relayNotice] : []),
-        ...(await loadCliApiStatusLines({ apiMode })),
-      ].join('\n'),
-    );
-  } catch (error: unknown) {
-    appendLocalAssistantTranscript(toErrorMessage(error));
-  }
-}
-
-const YOLO_USAGE = 'Usage: /yolo [ask | never | yolo]';
-
-function applyCliApprovalPolicySelection(
-  input: string,
-  context: SlashCommandContext,
-  usage = YOLO_USAGE,
-): void {
-  const normalized = input.trim().toLowerCase();
-  if (!normalized || normalized === 'status') {
-    openCliSlashCommandForm('approval', '');
-    return;
-  }
-
-  const policy = parseCliApprovalPolicy(normalized);
-  if (!policy) {
-    appendLocalAssistantTranscript(usage);
-    return;
-  }
-
-  context.setApprovalPolicy(policy);
-  appendLocalAssistantTranscript(
-    `Approval mode set to ${formatCliApprovalPolicy(policy)}.`,
-  );
-}
-
-async function showCliMemoryList(): Promise<void> {
-  appendLocalAssistantTranscript(formatCliMemoryList(await loadMemoryItems()));
-}
-
-export async function showCliMemoryPreview(inputPath: string): Promise<void> {
-  appendLocalAssistantTranscript(await formatCliMemoryPreview(inputPath));
-}
-
-function openCanonicalSlashForm(
-  commandName: string,
-  registered: SlashCommand | undefined,
-  remainder: string,
-): void {
-  if (registered && openRegisteredCliSlashForm(registered, remainder)) return;
-  openCliSlashCommandForm(commandName, remainder);
-}
+// Public surface preserved for `runChatTui` and the slash-command tests, which
+// import these from `handleSlashCommand` rather than the per-domain modules.
+export {
+  applyCliModelSelection,
+  applyInitialCliAgentSelection,
+  chatAgentSupportsDelegation,
+  chatToolUseAgentUsageError,
+} from './handlers/agentModelCommands';
+export { applyCliApiModeSelection } from './handlers/apiModeCommands';
+export { showCliMemoryPreview } from './handlers/memoryCommands';
+export {
+  CHAT_API_MODE_MODEL_RECOVERY,
+  type SlashCommandContext,
+} from './handlers/slashContext';
 
 export async function handleTuiSlashCommand(
   line: string,
@@ -567,7 +148,7 @@ export async function handleTuiSlashCommand(
       }
       return true;
     case 'yolo':
-      applyCliApprovalPolicySelection(rest || 'yolo', context, YOLO_USAGE);
+      applyCliApprovalPolicySelection(rest || 'yolo', context);
       return true;
     case 'status': {
       const meta = cliState.sessionMeta.get();
