+++ packages/cli/src/chat/tui/commands/handlers/apiModeCommands.ts
@@ -1,0 +1,102 @@
+import {
+  formatCliApiMode,
+  parseCliApiMode,
+  setCliApiMode,
+  type CliApiMode,
+} from '@cli/runtime/apiAccessMode';
+import { loadCliApiStatusLines } from '@cli/runtime/apiStatus';
+import { formatCliNoAvailableModelsRecovery } from '@cli/runtime/modelAccess';
+import { selectCliRootModel } from '@cli/runtime/rootModelSelection';
+
+import { cliState } from '@cli/chat/tui/state/cliState';
+import { chatTuiCanStartRootRun } from '@cli/chat/tui/state/sessionRunState';
+import { appendLocalAssistantTranscript } from '@cli/chat/tui/state/transcript';
+import { toErrorMessage } from '@common/errors/errorMessage';
+import {
+  CHAT_API_MODE_MODEL_RECOVERY,
+  type SlashCommandContext,
+} from './slashContext';
+
+async function reconcileRootModelAfterApiModeChange(
+  context: SlashCommandContext | undefined,
+  apiMode: CliApiMode,
+): Promise<string | undefined> {
+  if (!context || !chatTuiCanStartRootRun(context.session)) return undefined;
+
+  const { model: currentModel, modelSource } = cliState.sessionMeta.get();
+  const selection = await selectCliRootModel({
+    model: currentModel,
+    modelSource,
+    apiMode,
+    noAvailableModelsMessage: formatCliNoAvailableModelsRecovery(
+      apiMode,
+      CHAT_API_MODE_MODEL_RECOVERY,
+    ),
+  });
+  if (selection.model === currentModel) return undefined;
+
+  cliState.sessionMeta.set({
+    ...cliState.sessionMeta.get(),
+    model: selection.model,
+  });
+  return selection.notice;
+}
+
+/** Set the chat session's api-mode without touching the persisted global. */
+export function setCliSessionApiMode(apiMode: CliApiMode): void {
+  cliState.sessionMeta.set({
+    ...cliState.sessionMeta.get(),
+    apiMode,
+  });
+}
+
+export async function applyCliApiModeSelection(
+  mode: string | CliApiMode,
+  context?: SlashCommandContext,
+): Promise<void> {
+  const normalized = mode.trim().toLowerCase();
+
+  if (!normalized || normalized === 'status') {
+    const lines = await loadCliApiStatusLines({
+      apiMode: cliState.sessionMeta.get().apiMode,
+    });
+    appendLocalAssistantTranscript(
+      [...lines, 'Usage: /api personal | /api included'].join('\n'),
+    );
+    return;
+  }
+
+  const apiMode = parseCliApiMode(normalized);
+  if (apiMode) {
+    await setCliApiMode(apiMode);
+    setCliSessionApiMode(apiMode);
+    let modelNotice: string | undefined;
+    try {
+      modelNotice = await reconcileRootModelAfterApiModeChange(
+        context,
+        apiMode,
+      );
+    } catch (error: unknown) {
+      modelNotice = toErrorMessage(error);
+    }
+    appendLocalAssistantTranscript(
+      [
+        `API mode set to ${formatCliApiMode(apiMode)}.`,
+        ...(modelNotice ? [modelNotice] : []),
+      ].join('\n'),
+    );
+    return;
+  }
+
+  appendLocalAssistantTranscript('Usage: /api personal | /api included');
+}
+
+export async function showCliAuthStatus(): Promise<void> {
+  appendLocalAssistantTranscript(
+    (
+      await loadCliApiStatusLines({
+        apiMode: cliState.sessionMeta.get().apiMode,
+      })
+    ).join('\n'),
+  );
+}
